libsyntax: Unconfigure tests during normal build
